pressure solve working

+++ src/fluid_sim/forces.rs
@@ -2,29 +2,29 @@ use super::particle::Particle;
 use nalgebra as na;
 
 pub trait Force {
-    fn get_force(&self, particle: &Particle) -> na::Vector3<f32>;
+    fn get_force(&self, particle: &Particle) -> na::Vector3<f64>;
 }
 
 pub struct ConstantForce {
-    pub force: na::Vector3<f32>
+    pub force: na::Vector3<f64>
 }
 
 impl Force for ConstantForce {
-    fn get_force(&self, particle: &Particle) -> na::Vector3<f32> {
+    fn get_force(&self, particle: &Particle) -> na::Vector3<f64> {
         self.force
     }
 }
 
 
 pub struct SphereForce {
-    pub center: na::Vector3<f32>,
-    pub radius: f32,
-    pub force: na::Vector3<f32>,
+    pub center: na::Vector3<f64>,
+    pub radius: f64,
+    pub force: na::Vector3<f64>,
 }
 
 
 impl Force for SphereForce {
-    fn get_force(&self, particle: &Particle) -> na::Vector3<f32> {
+    fn get_force(&self, particle: &Particle) -> na::Vector3<f64> {
         if (particle.position - self.center).norm().abs() < self.radius {
             self.force
         } else {
